refactor storage into multiple files
