+++ src/index.js
@@ -1,3 +1,3 @@
-import storage from "./storage";
+import storage from "./storage/storage";
 
-export {storage};
+export { storage };
